feat: 메인 및 혼자정하기 페이지 구조 생성 및 metadata 추가
