+++ src/app/page.tsx
@@ -1,72 +1,18 @@
-'use client';
+import type { Metadata } from 'next';
+import MainClient from './MainClient';
 
-import { useQuery } from '@tanstack/react-query';
-
-import Carousel from '@/shared/components/Carousel';
-import type { CarouselItem } from '@/shared/components/Carousel';
-import Clock from '@/shared/components/Clock/Clock';
-import WeatherMood from '@/domain/WeatherMood/WeatherMood';
-import RoomEntryCard from '@/domain/Room/RoomEntryCard/RoomEntryCard';
-import QuoteCard from '@/domain/QuoteCard';
-
-import { menusServiceClient } from '@/app/services/backend/menus.api';
-import { mapTopFavoritesToCarousel } from '@/app/utils/mapTopFavoritesToCarousel';
-
-import styles from './page.module.scss';
-
-// 서버 없을 때도 보여줄 fallback
-const FALLBACK_ITEMS: CarouselItem[] = [
-  {
-    id: 'fallback-1',
-    menuName: '추천 메뉴 준비중',
-    rank: 1,
-    favoriteCount: 0,
-    stores: [],
+export const metadata: Metadata = {
+  title: '오늘 뭐 먹지? | 직장인 점심 메뉴 추천 & 맛집 지도',
+  description:
+    '매일 반복되는 점심 고민, 룰렛으로 해결하세요. 날씨와 기분에 딱 맞는 메뉴 추천부터 근처 맛집 검색까지 한 번에!',
+  openGraph: {
+    title: '오늘 뭐 먹지? | 직장인 점심 메뉴 추천 & 맛집 지도',
+    description:
+      '매일 반복되는 점심 고민, 룰렛으로 해결하세요. 날씨와 기분에 딱 맞는 메뉴 추천부터 근처 맛집 검색까지 한 번에!',
+    url: 'https://whatlunch.vercel.app',
   },
-  {
-    id: 'fallback-2',
-    menuName: '추천 메뉴 준비중',
-    rank: 2,
-    favoriteCount: 0,
-    stores: [],
-  },
-  {
-    id: 'fallback-3',
-    menuName: '추천 메뉴 준비중',
-    rank: 3,
-    favoriteCount: 0,
-    stores: [],
-  },
-];
+};
 
 export default function HomePage() {
-  const {
-    data: topMenus,
-    isLoading,
-    isError,
-  } = useQuery({
-    queryKey: ['menus', 'favorites', 'top'],
-    queryFn: () => menusServiceClient.getTopFavoriteMenus(3),
-    staleTime: 30_000,
-  });
-
-  const carouselItems: CarouselItem[] =
-    !isLoading && !isError && topMenus?.length
-      ? mapTopFavoritesToCarousel(topMenus)
-      : FALLBACK_ITEMS;
-
-  return (
-    <div className={styles['container']}>
-      <div className={styles['container__left']}>
-        <Carousel duration={4000} items={carouselItems} />
-        <RoomEntryCard />
-      </div>
-
-      <div className={styles['container__right']}>
-        <Clock />
-        <WeatherMood />
-        <QuoteCard />
-      </div>
-    </div>
-  );
+  return <MainClient />;
 }
