+++ src/styles/base/_global.scss
@@ -40,3 +40,16 @@ a {
     color-scheme: dark;
   }
 }
+
+// 시각적으로 숨기지만 스크린 리더에서는 읽히는 텍스트
+.sr-only {
+  position: absolute;
+  width: 1px;
+  height: 1px;
+  padding: 0;
+  margin: -1px;
+  overflow: hidden;
+  clip: rect(0, 0, 0, 0);
+  white-space: nowrap;
+  border-width: 0;
+}
